melhoramento dos coeficientes a e b

+++ MMQ.py
@@ -27,8 +27,8 @@ class MMQ:
         self.lnxlny = [self.lnx[i]*self.lny[i] for i in range(self.n)]
         self.slnxlny = sum(self.lnxlny)
     def lin(self):
-        a = ((self.n*self.sxy)-(self.sx*self.sy))/((self.n*self.sx2)-(self.sx)**2)
-        b = ((self.sxy*self.sx)-(self.sy*self.sx2))/((self.sx**2)-(self.n*self.sx2))
+        a = float(((self.n*self.sxy)-(self.sx*self.sy))/((self.n*self.sx2)-(self.sx)**2))
+        b = float(((self.sxy*self.sx)-(self.sy*self.sx2))/((self.sx**2)-(self.n*self.sx2)))
         return f'{a}*x {"+" if b >= 0 else "-"} {abs(b)}'
     def graf_lin(self):
         ypnt = [resolva(self.lin(), i) for i in self.x]
@@ -37,8 +37,8 @@ class MMQ:
         plt.ylabel('concentração de CO2 em ppm'); plt.xlabel('ano')
         plt.grid(); plt.show()
     def log(self):
-        a = ((self.n*self.sylnx)-(self.slnx*self.sy))/((self.n*self.sln2x)-(self.slnx)**2)
-        b = ((self.sylnx*self.slnx)-(self.sy*self.sln2x))/((self.slnx**2)-(self.n*self.sln2x))
+        a = float(((self.n*self.sylnx)-(self.slnx*self.sy))/((self.n*self.sln2x)-(self.slnx)**2))
+        b = float(((self.sylnx*self.slnx)-(self.sy*self.sln2x))/((self.slnx**2)-(self.n*self.sln2x)))
         return f'{a}*ln(x) {"+" if b >= 0 else "-"} {abs(b)}'
     def graf_log(self):
         ypnt = [resolva(self.log(), i) for i in self.x]
@@ -47,9 +47,9 @@ class MMQ:
         plt.ylabel('concentração de CO2 em ppm'); plt.xlabel('ano')
         plt.grid(); plt.show()
     def exp(self):
-        a = ((self.n*self.sxlny) - (self.sx*self.slny))/(self.n*self.sx2-(self.sx**2))
-        b = ((self.sxlny*self.sx) - (self.slny*self.sx2))/((self.sx**2)-self.n*self.sx2)
-        eb = exp(b)
+        a = float(((self.n*self.sxlny) - (self.sx*self.slny))/(self.n*self.sx2-(self.sx**2)))
+        b = float(((self.sxlny*self.sx) - (self.slny*self.sx2))/((self.sx**2)-self.n*self.sx2))
+        eb = float(exp(b))
         return f'{eb}*exp({a}*x)'
     def graf_exp(self):
         ypnt = [resolva(self.exp(), i) for i in self.x]
@@ -58,9 +58,9 @@ class MMQ:
         plt.ylabel('concentração de CO2 em ppm'); plt.xlabel('ano')
         plt.grid(); plt.show()
     def pot(self):
-        a = ((self.n*self.slnxlny)-(self.slnx*self.slny))/((self.n*self.sln2x)-(self.slnx**2))
-        b = ((self.slnx*self.slnxlny)-(self.slny*self.sln2x))/((self.slnx**2)-self.n*self.sln2x)
-        eb = exp(b)
+        a = float(((self.n*self.slnxlny)-(self.slnx*self.slny))/((self.n*self.sln2x)-(self.slnx**2)))
+        b = float(((self.slnx*self.slnxlny)-(self.slny*self.sln2x))/((self.slnx**2)-self.n*self.sln2x))
+        eb = float(exp(b))
         return f'{eb}*x^{a}'
     def graf_pot(self):
         ypnt = [resolva(self.pot(), i) for i in self.x]
